AnswerId should be uint

+++ NetCord/MessagePollAnswerProperties.cs
@@ -5,7 +5,7 @@ namespace NetCord;
 public partial class MessagePollAnswerProperties(MessagePollMediaProperties pollMedia)
 {
     [JsonPropertyName("answer_id")]
-    public ulong AnswerId { get; set; }
+    public uint AnswerId { get; set; }
     
     [JsonPropertyName("poll_media")]
     public MessagePollMediaProperties PollMedia { get; set; } = pollMedia;
